dayEqualSaturday
Value of the for loop variable used for the condition to create a new line on Saturday, the last day of the week. Since Saturday has a placeholder of 6 on the daysOfWeek array ,if the value of the day counter (i) is equivalent to daysOfWeek[6] or "S", a new line is preinted.